requirements.txt
streamlit
pypdf2
langchain
langchain-ollama
sentence-transformers
faiss-cpu
ollama
